requirements.txt
openai>=1.54.0
pillow>=10.0.0
python-dotenv>=1.0.0
